requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
Async-OAuthlib==0.0.9
async-timeout==3.0.1
attrs==21.2.0
blinker==1.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
discord-ext-ipc==2.1.1
discord.py==1.6.0
distest==0.6.2
docopt==0.6.2
Flask==2.0.2
Flask-Discord==0.1.64
gunicorn==20.1.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.0
hyperframe==6.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
oauthlib==3.1.1
prettytable==2.4.0
priority==2.0.0
PyJWT==2.3.0
python-dotenv==0.19.2
quart==0.16.2
Quart-Discord==2.1.4
requests==2.26.0
requests-oauthlib==1.3.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wsproto==1.0.0
yarg==0.1.9
yarl==1.7.2
